# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "A RESTful API for a news aggregator service that pulls articles from various sources and provides endpoints for a frontend application to consume.",
        "contact": {
            "email": "stefan.prence@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/v1/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get all articles",
                "operationId": "024de9182405b8b7593efbdf02a7ba30",
                "responses": {
                    "200": {
                        "description": "A list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/articles/search": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Search for articles",
                "operationId": "393ee2b7b643baa08706ed2b23e82be2",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/keyword"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/from_date"
                    },
                    {
                        "$ref": "#/components/parameters/to_date"
                    },
                    {
                        "$ref": "#/components/parameters/source"
                    },
                    {
                        "$ref": "#/components/parameters/author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Validation error details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/articles/{article}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get article by ID",
                "operationId": "843d3068e2912fa5e4253283da3fe319",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/news-feed": {
            "get": {
                "tags": [
                    "User News Feed"
                ],
                "summary": "Get user news-feed",
                "operationId": "fbea16dbdebe2b805ac5090f2869dffa",
                "responses": {
                    "200": {
                        "description": "A list of user articles based on preferences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "operationId": "ca6771fbbfe8c70e07c9c92355cbb7f9",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of user preferences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Preference"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Add user preference",
                "operationId": "990ed640af19b272bc35ff8d28222e6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preference_type": {
                                        "type": "string",
                                        "example": "source"
                                    },
                                    "preference_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": "[1,2,3]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A list of recently added user preferences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Preference"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Validation error details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Article": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "content": {
                                    "type": "string"
                                },
                                "authors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                },
                                "url": {
                                    "type": "string",
                                    "format": "url"
                                },
                                "published_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "source": {
                                    "$ref": "#/components/schemas/Source"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Author": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Source": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Preference": {
                "required": [
                    "id",
                    "user_id",
                    "preferable_id",
                    "preferable_type"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 14
                                },
                                "preferable_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "preferable_type": {
                                    "type": "string",
                                    "example": "Source"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "keyword": {
                "name": "keyword",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "example": 10
                }
            },
            "from_date": {
                "name": "from_date",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-10-01"
                }
            },
            "to_date": {
                "name": "to_date",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-10-31"
                }
            },
            "source": {
                "name": "source",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "example": "The Guardian"
                }
            },
            "author": {
                "name": "author",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}